Separate controller package 1. separate controller 2. Swagger 3. Added application context 4. http://localhost:8080/test/swagger-ui.html#/test45api 5. http://localhost:8080/test/
