+++ pom.xml
@@ -21,8 +21,21 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
-</dependencies>
-<build>
+		
+	<!-- swagger2 dependency-->	
+		<dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-swagger2</artifactId>
+			<version>2.7.0</version>
+		</dependency>
+		<dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-swagger-ui</artifactId>
+			<version>2.7.0</version>
+		</dependency>
+			
+	</dependencies>
+		<build>
 		<plugins>
 			<plugin><!-- Include if you want to make an executable jar[FAT JAR which 
 					includes all dependencies along with sprinboot loader] that you can run on
