+++ src/main/java/com/thomascook/testapi/swagger/SwaggerConfig.java
@@ -1,0 +1,28 @@
+package com.thomascook.testapi.swagger;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+
+@ComponentScan(basePackages = {"com.thomascook.testapi.controller"})
+// SwaggerConfig
+@Configuration
+@EnableSwagger2
+public class SwaggerConfig {   
+	// http://localhost:8080/swagger-ui.html
+    @Bean
+    public Docket api() { 
+        return new Docket(DocumentationType.SWAGGER_2)  
+          .select()                                  
+          .apis(RequestHandlerSelectors.any())              
+          .paths(PathSelectors.any())                          
+          .build();                                           
+    }
+}
